Added the menu screens that I am doing and classes that everyone is doing
I have added Java files for each of the classes we are doing into a package called "AppDataStructures". Also, I have added the Java and XML files for each of the menu screens that I am doing (the Options menu, Session Recording Menu, and the Start/End Session Menu).

Hope this commit works correctly and everything

- Aaron Batch