Added more project files and worked on my views
I have added blank activity files (Java and XML files) for the "Session Information Menu" and the "Session Viewing Menu". So whoever is working on those can now start on those activities. I have not yet added the files for the activities that Naumaan is working on (Main menu and Weekly Report Menu) since I do not currently know if he has created those files already in his version of the project and just hasnt commited them yet.

Also, I have worked a bit on the views that I am doing (the Session Recording Menu and the Start/End Session menu). I have not started work on the Options menu yet. Currently, you can click on each type of session in the session recording menu, and a string of the specific type of session will be sent to the Start/End Session menu (so that the Start/End session menu knows which kind of activity the user wants to start). Also, the back buttons on both views work correctly.