Implemented and tested Child and Session class
I have implemented and tested both the Child class and Session superclass. Each class comes with the data and functions that are needed for both the Child, as well as the Session superclass. The Child has data such as a first name, last name, date of birth, and other things. The Session superclass has various data in it such as a start time, end time, and other things. Be sure to look at each class before using it to learn of what other data and functions it has. Also, I had to take off the "extends Session" line of each session subclass off in order to commit a project that is runnable. So before getting these changes, be sure to  back up your class (just in case something weird happens, even though I dont think anything bad should happen).

- Aaron Batch