Implemented a singleton in the "Child" class
I have implemented a Singleton inside of the class "Child". This means that there will be only one instance of a child existing throughout the entire application (kind of like a child global variable that can be read and updated by all of the activities in the application in real time). To get the child instance so you can use it in your class, just call the "Child.getChild()" function which will return the instance. After this, you can edit and get information from the child. For more information, just look inside of the "Child" class.