Implemented more of the StartOrEndSessionMenu
I implemented most of the StartOrEndSessionMenu, however, I just realized that on all of the lines where I compare if a string is equal to saomething, I have been saying "if( string == "blah")". Apparently, that is not the correct way to do that. Instead you use "if(string.isEqual("blah"))". I will have to change that in my code.

Also, I made the TimeVars enum class its own file