Tested and finished the SessionRecordingMenu, and tested and finished most of the StartOrEndSessionMenu
I have tested and finished the session recording menu. Also, I have tested and finished the start/end session menu when the user is coming from the session recording menu. I still however need to test it when the user is coming from the session viewing menu, but we need a main menu for that and I need to check with Nathan on his status on the Session Viewing menu. The menu should however work with the session viewing menu, but we still need to test it. Also, I noticed that I forgot to instantiate the child singleton's array list with an empty array, so I also did that.