Changed the Child singleton and added some variables to the Settings singleton
I have changed the Child singleton to work for multiple children. Basically, I have turned the child singleton into a child array list singleton. The way it is structured should allow us to not have to refactor a lot of our existing code. For example, the getChild() function is structured the same way (it returns a child object), but now it returns a child object at a certain index from the singleton array list instead of just the same child (like last time).

Along with that, I also added in some static functions in the child object that allow you to do various things (such as add a child, remove a child, get the child arraylist, etc.). In order to combat app-breaking bugs, I also added in some (potentially redundant) code into the functions which check for certain error cases (but it is not perfect though).

Lastly, I also added the "current child index" and "current number of children"  variables to the settings object.