Share Weekly Report Menu Update
Hello,

I've went ahead and done the Share Weekly Report Menu. It pulls the information in a similar style to the Information Viewing Session Menu, but lists out the addition settings that are stored in the respective appDataStructures.

I've tested in and it works fully (to my knowledge) with the Feeding Session menu. For the other three, it just prints out blanks since their Session information menu isn't available to pass information.

Edited the activity_share_weekly_report_menu.xml to provide a basic menu template and put in a share button, that for now only share some basic text. Purpose in the future is the actually pass all the information from the different sessions into one string file and then pass that to this "share" button.

In colors, included a simple white color.