Finished the sleeping info session and added some other stuff
I have finished the sleeping information menu, and have also added a time picker fragment as well as a date picker fragment to the project. The nice thing about these picker fragments is that they are components that function by themselves, which means that they can be added to any activity that needs them (as long as that activity implements the methods that it needs).

Also, I have changed the addChild method so that instead of taking a child object as input, it actually takes all of the info needed for a child object, and then makes a child inside the function. So Nathan should be able to test his view with this (until I get the Add child menu finished).