I changed the Child singleton a bit
I added a function called "setUpChildren" which initializes the child arraylist singleton with an empty array list of children (rather than placing a mock child into the array like the getChild function does). Also, in the event that the array list singleton is null, the getChildren function will call setUpChildren and return an empty array list.