I have implemented the Add Child Menu
I have finished and tested the add child menu. Also, I edited the date picker fragment a bit to fit in a style variable (to change the look of the date picker). The variable to change is in the settings singleton called the "styleInt". For some reason however, the date picker is not changing to the date picker that I want for the add child menu. However, it is changing to a valid date picker (just not the one I want). But for right now, I will live with it (and change it in the future), because it is currently functional.