Added a sign in menu for DropBox
I have added a sign in menu for Dropbox. The sign in menu is now the first view that loads on app startup. It is kind of annoying because each time you reload the application, you have to connect to your Dropbox accout, but such is life at the moment. As a hack, I added a button under the sign in button that just sends you to the choose child menu. However, I dont think you will be connected to Dropbox if you go that route, so it may not see your changes. We can take that hack button out once we figure out how to get around this. I only added that button for easier testing purposes.